make a data storage class
Before I was loading all the species and bootstrap iterations into
a job list to iterate over. This was fine with the small harvard
dataset, but was using > 5GB of memory on only the master node
for the NPN dataset, causing issues with running on the hipergator.
The new data_store class dynamically loads new jobs of species and
bootstrapped data, which keeps the memory on the master node < 500mb.

TODO: Right now it only works with 1 dataset at a time. Make it able
to load all datasets to iterate over