refactor how doy estimates are done
before I would calculate the estimated doy for each unique site/year,
then for each observation match it up to it's respective site/year.
Now I gather the temperature data each observation and calculate
doy directly, negating the need for loop in calculate_doy_estimates,
and getting rid of calculating_doy_estimates altogether.
This was supposed to be faster because I got rid of all the for
loops, but that's not neccecarily the case. Prelim testing
on the harvard acer rubrum spp shows it was faster *with*
the old implimentation.