Scricter npn filtering (#20)
* put in stricter filtering

new option to enforce a maximum number of days that prior observations of status==0 can be made. to be used for NPN data only

also cleaned things up just a bit

* much scricter NPN requirements

from Crimmins et al. 2017 methods

-Date of first "yes" observation which was preceded by a "no" within 15 days prior
-More than one individual at a site? Take the mean value
-cutoff at doy 172 for buds and leaves
-cutoff at doy 213 for flrs
-minimum 30 observations at the end

* correctly get the prior doy observation

I took this out thinking it was redundant, but it's not. Before it was getting the  *first* observation of the year, but this way correctly gets the most prior observation to the first "yes"

* add some extra spp for study comparison

* round final values to nearest whole doy

* make the cutoff for the prior 0 observation to 30

this allows more samples across the different datasets to be compared